# Conflicts:
#	app/src/main/java/com/example/meetup/repository/AppDatabase.java
#	app/src/main/java/com/example/meetup/repository/dao/EventDao.java
#	app/src/main/java/com/example/meetup/view/home/HomeActivity.java
#	app/src/main/java/com/example/meetup/view/home/SplashActivity.java
#	app/src/main/java/com/example/meetup/view/home/event/EventsRepository.java

+++ app/src/main/java/com/example/meetup/services/LoadPersonalWorker.java
@@ -1,14 +1,21 @@
 package com.example.meetup.services;
 
 import android.content.Context;
+import android.content.SharedPreferences;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.lifecycle.ViewModelProvider;
 import androidx.work.Worker;
 import androidx.work.WorkerParameters;
 
+import com.example.meetup.R;
 import com.example.meetup.model.response.EventGetFromApi;
 import com.example.meetup.model.response.EventResponse;
 import com.example.meetup.repository.PersonalJoinedRepository;
+import com.example.meetup.ulti.Define;
+import com.example.meetup.ulti.MyApplication;
+import com.example.meetup.view.registerlogin.login.LoginViewModel;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -18,6 +25,11 @@ import retrofit2.Callback;
 import retrofit2.Response;
 
 public class LoadPersonalWorker extends Worker {
+
+    public SharedPreferences sharedPref = MyApplication.getAppContext()
+            .getSharedPreferences("tokenPref", Context.MODE_PRIVATE);
+    String token = sharedPref.getString("token",null);
+
     ApiUtils apiUtils = new ApiUtils();
     private EventJoinedServices mEventJoinedServices;
     public LoadPersonalWorker(@NonNull Context context, @NonNull WorkerParameters workerParams) {
@@ -27,33 +39,71 @@ public class LoadPersonalWorker extends Worker {
     @NonNull
     @Override
     public Result doWork() {
-        loadEventJoinedFromApi("eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9tZWV0dXAucmlra2VpLm9yZ1wvYXBpXC92MFwvbG9naW4iLCJpYXQiOjE1OTcyMDIzODQsImV4cCI6MTU5NzIwOTU4NCwibmJmIjoxNTk3MjAyMzg0LCJqdGkiOiI4bFh5dndJbFp1NmdGaXlWIiwic3ViIjo0NDUsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.N1PMDXTkkYQAGL0CnNCUDpHr_QbfDpn6ohk4E66Mb4w",1);
+        loadEventJoinedFromApi(token,Define.STATUS_WENT);
+        loadEventCanJoinFromApi(token,Define.STATUS_GOING);
         return Result.success();
     }
 
-    private void loadEventJoinedFromApi(String token, final long status) {
+    private void loadEventJoinedFromApi(final String token, final long status) {
     mEventJoinedServices = apiUtils.getEventJoinedServices();
     mEventJoinedServices.getListMyEventsJoined(token,status).enqueue(new Callback<EventResponse>() {
         @Override
         public void onResponse(Call<EventResponse> call, Response<EventResponse> response) {
             if (response.isSuccessful()){
-                List<EventGetFromApi> listEvent = response.body().getResponse().getEvents();
-                ArrayList listIdJoined = new ArrayList();
-                for (EventGetFromApi e: listEvent){
-                    listIdJoined.add(e.getId());
+                if(response.body().getStatus() == 0 ){
+                    sharedPref.edit().clear();
+                    sharedPref.edit().apply();
+                }else {
+                    List<EventGetFromApi> listEvent = response.body().getResponse().getEvents();
+                    List<Integer> listIdJoined = new ArrayList<>();
+                    for (EventGetFromApi e: listEvent){
+                        listIdJoined.add(e.getId());
+                    }
+                    PersonalJoinedRepository personalJoinedRepository = PersonalJoinedRepository.getInstance();
+                    personalJoinedRepository.deleteUsersEvents();
+                    personalJoinedRepository.updateUsersEvents(listIdJoined,status);
+                    personalJoinedRepository.getListEventJoined(Define.PAGE_SIZE_DEFAULT,status);
                 }
-                PersonalJoinedRepository personalJoinedRepository = PersonalJoinedRepository.getInstance();
-                personalJoinedRepository.getEventStatus(listIdJoined);
-                personalJoinedRepository.deleteUsersEvents();
-                personalJoinedRepository.updateUsersEvents(listIdJoined,status);
             }
         }
 
         @Override
         public void onFailure(Call<EventResponse> call, Throwable t) {
-
+            Toast.makeText(MyApplication.getAppContext(), R.string.api_error, Toast.LENGTH_LONG);
         }
     });
+
+    }
+
+
+    private void loadEventCanJoinFromApi(String token, final long status) {
+        mEventJoinedServices = apiUtils.getEventJoinedServices();
+        mEventJoinedServices.getListMyEventsJoined(token,status).enqueue(new Callback<EventResponse>() {
+            @Override
+            public void onResponse(Call<EventResponse> call, Response<EventResponse> response) {
+                if (response.isSuccessful()){
+                    if(response.body().getStatus() == 0 ){
+                        sharedPref.edit().clear();
+                        sharedPref.edit().apply();
+                    }else {
+                        List<EventGetFromApi> listEvent = response.body().getResponse().getEvents();
+                        ArrayList listIdCanJoin = new ArrayList();
+                        for (EventGetFromApi e : listEvent) {
+                            listIdCanJoin.add(e.getId());
+                        }
+                        PersonalJoinedRepository personalJoinedRepository = PersonalJoinedRepository.getInstance();
+                        personalJoinedRepository.deleteUsersEvents();
+                        personalJoinedRepository.updateUsersEvents(listIdCanJoin, status);
+                        personalJoinedRepository.getListEventJoined(Define.PAGE_SIZE_DEFAULT, status);
+                    }
+                }
+            }
+
+            @Override
+            public void onFailure(Call<EventResponse> call, Throwable t) {
+                Toast.makeText(MyApplication.getAppContext(), R.string.api_error, Toast.LENGTH_LONG);
+            }
+        });
 
     }
 }
